Rearrange MovieSelectionSlim
Moved service name & duration of current focus item so it doesn't
overlap with button hints. Mover the "space available" display so
that it doesn't overlap the selection list.

Mover the sort order and "after end of play" icons from off-screen
to on-screen.

Moved the synopsis of the current item to on-screen and reduced its
height,

Reduced the height of the selection list to accommodate the other
changes in the same screen height.

The MovieSelectionSlim screen selection list is now a bit tight if
the default 20 rows are displayed. It works well if the number of
rows is reduced to 15.